backend/requirements.txt
flask==3.1.2
flask-cors==6.0.1
pillow==12.0.0
pillow-heif==1.1.1
gunicorn==23.0.0
